::Provision backends expect user/pass prefixed by api_
The idea is to _not_ confuse the user: which username is that?
the one the user logs in the service, the username that is going
to be created on the server? Oh, it's the *API* username.
Password? The one I log on the service with, or the root
password for my new server? Oh, I see.. the *API* password.